// src/Models/CompaniaGD.ts
// src/Models/CompañiaGD.ts                          creadas por el admin general
export interface CompaniaGD {
  Id?: string;
  Title: string;                  // Nombre de la compañia (ESTUDIO DE MODA S A S, PARALELO S A S, STYLE INK SAS, DENIM HEAD SAS, EXPENSIS S.A.S, DH RETAIL SAS, MOVIMIENTO VISUAL SAS, META GRAPHICS SAS)
  FechaCreacion?: string;
  AdministradorCom?: string;         // Fecha de creación
  Activa?: boolean;
}